% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasnapshot.R
\name{data_snapshot}
\alias{data_snapshot}
\title{Data Snapshot for Two Numeric Variables}
\usage{
data_snapshot(data, x, y)
}
\arguments{
\item{`data`}{needs to be a data frame (for e.g., a tibble)}

\item{`x`}{This input is a numeric variable from a tibble or dataframe, for e.g. column with numeric values}

\item{`y`}{This input is a numeric variable from a tibble or dataframe, for e.g. column with numeric values}
}
\value{
This function returns summary tables for numeric variables, a scatterplot and adjusted R squared value
}
\description{
Function provides snapshot of the linear relationship between two numeric variables, as well as some
summary statistics for numeric variables.
}
\details{
When given two numeric variables, x and y, this function produces summary statistics and a plot
that allows the user to get a quick snapshot of the data and in particular their linear relationship
(if any).
}
\examples{
data_snapshot(mtcars, hp, mpg)
data_snapshot(gapminder, lifeExp, gdpPercap)

}
